+++ fbpcs/private_computation_cli/private_computation_cli.py
@@ -16,7 +16,6 @@ Usage:
     pc-cli run_stage <instance_id> --stage=<stage> --config=<config_file> [--server_ips=<server_ips> --dry_run] [options]
     pc-cli get_instance <instance_id> --config=<config_file> [options]
     pc-cli get_server_ips <instance_id> --config=<config_file> [options]
-    pc-cli get_mpc <instance_id> --config=<config_file> [options]
     pc-cli run_study <study_id> --config=<config_file> --objective_ids=<objective_ids> --input_paths=<input_paths> [--output_dir=<output_dir> --tries_per_stage=<tries_per_stage> --result_visibility=<result_visibility> --run_id=<run_id> --graphapi_version=<graphapi_version> --graphapi_domain=<graphapi_domain> --dry_run] [options]
     pc-cli pre_validate [<study_id>] --config=<config_file> [--objective_ids=<objective_ids>] --input_paths=<input_paths> [--tries_per_stage=<tries_per_stage> --dry_run] [options]
     pc-cli cancel_current_stage <instance_id> --config=<config_file> [options]
@@ -86,7 +85,6 @@ from fbpcs.private_computation_cli.private_computation_service_wrapper import (
     cancel_current_stage,
     create_instance,
     get_instance,
-    get_mpc,
     get_server_ips,
     print_current_status,
     print_instance,
@@ -183,7 +181,6 @@ def main(argv: Optional[List[str]] = None) -> None:
             "run_stage": bool,
             "get_instance": bool,
             "get_server_ips": bool,
-            "get_mpc": bool,
             "run_study": bool,
             "pre_validate": bool,
             "run_attribution": bool,
@@ -380,9 +377,6 @@ def main(argv: Optional[List[str]] = None) -> None:
         logger.info(instance)
     elif arguments["get_server_ips"]:
         get_server_ips(config, instance_id, logger)
-    elif arguments["get_mpc"]:
-        logger.info(f"Get MPC instance: {instance_id}")
-        get_mpc(config, instance_id, logger)
     elif arguments["validate"]:
         logger.info(f"Validate instance: {instance_id}")
         validate(