+++ fbpcs/private_computation_cli/tests/test_private_computation_cli.py
@@ -286,16 +286,6 @@ class TestPrivateComputationCli(TestCase):
         pc_cli.main(argv)
         get_ips_mock.assert_called_once()
 
-    @patch("fbpcs.private_computation_cli.private_computation_cli.get_mpc")
-    def test_get_mpc(self, get_mpc_mock) -> None:
-        argv = [
-            "get_mpc",
-            "instance123",
-            f"--config={self.temp_filename}",
-        ]
-        pc_cli.main(argv)
-        get_mpc_mock.assert_called_once()
-
     @patch("fbpcs.private_computation_cli.private_computation_cli.TokenValidator")
     @patch("fbpcs.private_computation_cli.private_computation_cli.BoltGraphAPIClient")
     @patch("fbpcs.private_computation_cli.private_computation_cli.run_study")